+++ src/types/breadcrumbType.ts
@@ -1,0 +1,5 @@
+export interface BreadcrumbItem {
+  label: string;
+  href: string;
+  isCurrent?: boolean;
+}
